restructured modals in events (fix that delete modal visual bug)
